Move all of the Angular activation into the Dockerfile-postgenerate (#16)
