+++ hooks/post_gen_project.py
@@ -66,7 +66,10 @@ def generate_backbone_translations():
     return True
 
 
-bootstrap_subprojects = make_bootstrap_command('bootstrap')
+bootstrap_subprojects = Command(
+    'Finalize subproject package configuration',
+    ['docker', 'compose', '-f', 'compose-postgenerate.yml', 'up']
+)
 
 cd_into_project = Command('', ['cd', SLUG])
