+++ {{cookiecutter.slug}}/frontend.angular/Dockerfile-postgenerate
@@ -1,4 +1,4 @@
-FROM node:18 as bootstrap
+FROM node:18 AS bootstrap-1
 
 WORKDIR /usr/src/app
 RUN yarn init -yp && \
@@ -15,25 +15,33 @@ RUN yarn init -yp && \
 WORKDIR /usr/src/app/{{cookiecutter.npm_slug}}
 COPY . .
 
-VOLUME /usr/src/app/frontend
-CMD cp package.json package.overwrite.json ../frontend/
+FROM python:3.9 AS bootstrap-2
 
-FROM bootstrap as bootstrap-2
+WORKDIR /usr/src/app
+COPY --from=bootstrap-1 \
+     /usr/src/app/{{cookiecutter.npm_slug}} frontend/
 
-WORKDIR /usr/src/app/{{cookiecutter.npm_slug}}
-COPY package.json ./
+WORKDIR /usr/src/app/frontend
+RUN python bootstrap/override_json.py package
+
+FROM node:18 AS bootstrap-3
+
+WORKDIR /usr/src/app/frontend
 RUN rm src/favicon.ico .editorconfig && \
     yarn && \
     yarn ng add @angular/localize --skip-confirmation
 
-VOLUME /usr/src/app/frontend
-CMD cp angular.json angular.overwrite.json ../frontend/
+FROM python:3.9 AS bootstrap-4
 
-FROM bootstrap-2 as bootstrap-3
+WORDIR /usr/src/app/frontend
+RUN python bootstrap/override_json.py angular
 
-WORKDIR /usr/src/app/{{cookiecutter.npm_slug}}
-COPY angular.json ./
+FROM node:18 AS bootstrap-5
+
+WORKDIR /usr/src/app/frontend
 RUN yarn ng extract-i18n --output-path locale
 
-VOLUME /usr/src/app/frontend
-CMD cp -r * .* ../frontend/
+FROM python:3.9 AS bootstrap-6
+
+WORKDIR /usr/src/app/frontend
+RUN python bootstrap/finish_localizations.py
